Disable properties w/o proper handling
We can currently only properly filter string properties. Some non-string
props were appearing in the properties list, but they couldn't be
handled. This meant the "Add Property" button was behaving strangely.
Disabling these non-string properties resolves the issue, but in the
long-term we definitely need to add handling for other types, or
strictly enforce that all types are string.