Add read-only permissions for c2m2 data
Updates the load script for the c2m2 data to make sure a dedicated
database is created for the data (it was being written to the default
"neo4j" database previously). This db is read-only. Also creates a
special user for reading the c2m2 data, which only has minimal read
access to the database.

Together these changes should help eliminate the risk of a malicious
user attempting to modify the data or gain undesirable privileges. We
should still be cautious about the risk of cypher query injection from
the frontend, but this is a good first step towards preventing
users from causing mayhem in the database.