Fix cytoscape event handler binding bug
We should be careful using the `removeAllListeners` method of the
cytoscape core object in the future, as this will also remove any
listeners bound by layout packages. This was only apparent with the
d3-force layout, but it could apply to others as well and is a nasty
source of bugs.