Add pathway term search improvements
Fixes an issue where if more than two terms had the same synonym, all
but one was dropped from the list. Also fixes an issue where duplicates
of the same term were appearing in the list if the searched term matched
more than one synonym pointing to the same term.